Cleanup and fix reset of state by init
