+++ src/obr/OpenFOAM/BlockMesh.py
@@ -3,6 +3,8 @@
 import setFunctions as sf
 from core import modifies_file
 from typing import TYPE_CHECKING, Any
+from subprocess import check_output
+import sys
 
 
 if TYPE_CHECKING:
@@ -15,6 +17,61 @@ if TYPE_CHECKING:
     _Base = OpenFOAMCase
 else:
     _Base = object
+
+
+def calculate_simple_partition(nSubDomains, decomp):
+    """Calculates a simple domain decomposition based on nSubDomains
+
+    Returns
+    -------
+        number of subdomains
+
+    """
+
+    domains = lambda x: x[0] * x[1] * x[2]
+    remainder = lambda x: nSubDomains - domains(x)
+
+    def next_position(x):
+        if x[0] == x[1]:
+            return 0
+        if x[1] == x[2]:
+            return 1
+        return 2
+
+    def isPrime(n):
+        for i in range(2, n):
+            if n % i == 0:
+                return False, i  # i at this iteration is equal to the smallest factor
+        return True, n
+
+    i = next_position(decomp)
+    is_prime, factor = isPrime(nSubDomains)
+
+    if is_prime:
+        decomp[i] *= factor
+        return decomp
+    if nSubDomains > 0:
+        decomp[i] *= factor
+        return calculate_simple_partition(int(nSubDomains / factor), decomp)
+    else:
+        return decomp
+
+
+def sed(fn, in_reg_exp, out_reg_exp, inline=True):
+    """wrapper around sed"""
+    if sys.platform == "darwin":
+        ret = check_output(
+            ["sed", "-i", "", "s/" + in_reg_exp + "/" + out_reg_exp + "/g", fn]
+        )
+    else:
+        ret = check_output(
+            ["sed", "-i", "s/" + in_reg_exp + "/" + out_reg_exp + "/g", fn]
+        )
+
+
+def set_cells(blockMeshDict, old_cells, new_cells):
+    """ """
+    sed(blockMeshDict, old_cells, new_cells)
 
 
 # TODO use FileParse as a base for modifying the blockMeshDict
@@ -62,7 +119,7 @@ class BlockMesh(_Base):
 
         for block in blocks:
             orig_block, target_block = block.split("->")
-            sf.set_cells(
+            set_cells(
                 self.blockMeshDict,
                 orig_block,
                 target_block,
